Silence a MSVC warning
git-svn-id: https://svn.eduke32.com/eduke32@3991 1a8010ca-5511-0410-912e-c29ae57300e0

+++ polymer/eduke32/source/game.c
@@ -4362,7 +4362,7 @@ void G_DrawRooms(int32_t snum, int32_t smoothratio)
 
     {
         int32_t vis = p->visibility;
-        g_visibility = (vis <= 0) ? 0 : vis * (numplayers > 1 ? 1.f : r_ambientlightrecip);
+        g_visibility = (vis <= 0) ? 0 : (int32_t)(vis * (numplayers > 1 ? 1.f : r_ambientlightrecip));
     }
 
     CAMERA(sect) = p->cursectnum;
